Notifications, changed quantity logic

+++ app/src/main/java/com/ferenc/pamp/presentation/screens/main/chat/create_order/payment/select_card/SelectCardFragment.java
@@ -62,7 +62,7 @@ public class SelectCardFragment extends ContentFragment implements SelectCardCon
     protected TextView tvCardNumber;
 
     @FragmentArg
-    protected int mQuantity;
+    protected double mQuantity;
 
     @Bean
     protected UserRepository mUserRepository;
